Extracted CommandTestCase from TestGenerateStringFile.
