+++ test/test_generate_string_file.rb
@@ -1,79 +1,48 @@
-require 'twine_test_case'
+require 'command_test_case'
 
-class TestGenerateStringFile < TwineTestCase
-  def prepare_mock_formatter(formatter_class)
-    formatter = formatter_class.new(@mock_strings, {})
-    formatter_class.stubs(:new).returns(formatter)
+class TestGenerateStringFile < CommandTestCase
+  def new_runner(language, file)
+    options = {}
+    options[:output_path] = File.join(@output_dir, file) if file
+    options[:languages] = language if language
+
+    Twine::Runner.new(nil, options)
+  end
+
+  def prepare_mock_write_file_formatter(formatter_class)
+    formatter = prepare_mock_formatter(formatter_class)
     formatter.expects(:write_file)
   end
 
-  def setup
-    super
-
-    @known_languages = %w(en fr de sp)
-
-    @mock_strings = Twine::StringsFile.new
-    @mock_strings.language_codes.concat @known_languages
-    Twine::StringsFile.stubs(:new).returns(@mock_strings)
-
-    @mock_android_formatter = Twine::Formatters::Android.new(@mock_strings, {})
-    Twine::Formatters::Android.stubs(:new).returns(@mock_android_formatter)
-  end
-
   def test_deducts_android_format_from_output_path
-    options = {
-      output_path: File.join(@output_dir, 'fr.xml'),
-      languages: ['fr']
-    }
-    runner = Twine::Runner.new(nil, options)
+    prepare_mock_write_file_formatter Twine::Formatters::Android
 
-    @mock_android_formatter.expects(:write_file)
-
-    runner.generate_string_file
+    new_runner('fr', 'fr.xml').generate_string_file
   end
 
   def test_deducts_apple_format_from_output_path
-    prepare_mock_formatter Twine::Formatters::Apple
-    options = {
-      output_path: File.join(@output_dir, 'fr.strings'),
-      languages: ['fr']
-    }
-    runner = Twine::Runner.new(nil, options)
+    prepare_mock_write_file_formatter Twine::Formatters::Apple
 
-    runner.generate_string_file
+    new_runner('fr', 'fr.strings').generate_string_file
   end
 
   def test_deducts_jquery_format_from_output_path
-    prepare_mock_formatter Twine::Formatters::JQuery
-    options = {
-      output_path: File.join(@output_dir, 'fr.json'),
-      languages: ['fr']
-    }
-    runner = Twine::Runner.new(nil, options)
+    prepare_mock_write_file_formatter Twine::Formatters::JQuery
 
-    runner.generate_string_file
+    new_runner('fr', 'fr.json').generate_string_file
   end
 
   def test_deducts_gettext_format_from_output_path
-    prepare_mock_formatter Twine::Formatters::Gettext
-    options = {
-      output_path: File.join(@output_dir, 'fr.po'),
-      languages: ['fr']
-    }
-    runner = Twine::Runner.new(nil, options)
+    prepare_mock_write_file_formatter Twine::Formatters::Gettext
 
-    runner.generate_string_file
+    new_runner('fr', 'fr.po').generate_string_file
   end
 
   def test_deducts_language_from_output_path
     random_language = @known_languages.sample
-    options = {
-      output_path: File.join(@output_dir, "#{random_language}.xml"),
-    }
-    runner = Twine::Runner.new(nil, options)
+    formatter = prepare_mock_formatter Twine::Formatters::Android
+    formatter.expects(:write_file).with(anything, random_language)
 
-    @mock_android_formatter.expects(:write_file).with(anything, random_language)
-
-    runner.generate_string_file
+    new_runner(nil, "#{random_language}.xml").generate_string_file
   end
 end
